+++ shared/src/index.ts
@@ -1,5 +1,6 @@
 export * from './models/common';
 export * from './models/room';
 export * from './models/uno';
+export * from './games';
 export * from './events/game-events';
 export * from './protocol/messages';